Added signals, mutators, and an operator. Changed Node state semantic.
Added signals "activated" and "deactivated" that emits whenever the Node
state (m_active) is changed.
Added mutators "value" and "deactivate", to get a Node value (m_value),
and to change Node's state to deactivated.
Added operator== for comparison of Nodes by their values.
Now, the Node's state is changed by double clicking on the Node, instead
of clicking (onde) on it. Hence the removal of old mouse events, and
adding a new one (mouseDoubleClickEvent).